Move SystemInfo to its own cpp/h
